Submit updated Galaxy proof of concept workflow
This is essentially @lwratten's workflow, I've only updated Salmon to
version 1.5.1 and used a dataset collection as the input.

I've uploaded the workflow to
[Dockstore](https://dockstore.org/workflows/github.com/mvdbeek/bioinformatics-workflows/salmon-proof-of-concept-galaxy:master?tab=info),
which is where we will encourage our users to submit their best-practice
workflows to. Users can import the workflow from there into their
favorite Galaxy server.